pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "autograder"
version = "0.1.0"
description = "Automatic Handwritten Answer Sheet Evaluator"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.0.0",
    "reportlab>=4.0.0",
    "python-dotenv>=1.0.0",
    "pypdf>=4.0.0",
    "pymupdf>=1.24.0",
    "psycopg[binary]>=3.2.0",
    "pgvector>=0.3.0",
]
dev = ["pytest>=7.0.0", "pytest-cov>=4.0.0"]
postgres = [
    "psycopg[binary]>=3.2.0",
    "pgvector>=0.3.0",
]
ocr_google = ["google-cloud-vision>=3.4.0"]
ocr_azure = ["azure-ai-vision-imageanalysis>=1.0.0b1", "azure-core>=1.29.5"]
ocr_paddle = ["paddleocr>=2.7.0", "paddlepaddle>=2.5.0"]
ocr_trocr = ["torch>=2.0.0", "transformers>=4.34.0", "Pillow>=9.5.0"]
llm_gemini = ["google-generativeai>=0.8.0"]
all = [
    "fastapi>=0.100.0", "uvicorn>=0.22.0",
    "google-generativeai>=0.8.0",
    "google-cloud-vision>=3.4.0",
    "azure-ai-vision-imageanalysis>=1.0.0b1", "azure-core>=1.29.5",
    "paddleocr>=2.7.0", "paddlepaddle>=2.5.0",
    "torch>=2.0.0", "transformers>=4.34.0", "Pillow>=9.5.0",
    "psycopg[binary]>=3.2.0", "pgvector>=0.3.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"
